The changes in the `aicommit.sh` file are:
* A new line is added at the end of the file, indicating that this is a new file with mode 100755.
* The `DB_PGSQL_TEST.SH` file has been created as a new file with mode 100644.

The contents of the `aicommit.sh` file have been modified to include a new section that checks for disk space, RAM usage, and long-running PostgreSQL queries. The output is displayed in a format wrapped to prevent exceeding the maximum line length.

The `DB_PGSQL_TEST.SH` file contains a single line that echoes a message asking the user to enter the time in minutes for which they want to check the query runtime.